Add files via upload
The code which you are seeing is based on python 3.7 using if, elif & else conditions. First I have indicated the code for food items as it will be easy for the input provider and result in a less chance of mistakes , created a code to make a list of your ingredient code which will be inserted in the input box. Then made different lists of what food items that can be created with different ingredients code. Now I have coded to check whether the input list has any matching elements to the Lists made for various food items. If there is any match then it will provide you with a list of all possible meals you can prepare.